ec2-deployment/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
pandas==2.0.3
numpy==1.24.3 